check

+++ services/workers/src/test/java/org/sagebionetworks/table/worker/TableCSVDownloadWorkerIntegrationTest.java
@@ -1,7 +1,8 @@
 package org.sagebionetworks.table.worker;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
@@ -13,6 +14,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.junit.After;
+import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,10 +25,9 @@ import org.sagebionetworks.repo.manager.UserManager;
 import org.sagebionetworks.repo.manager.asynch.AsynchJobStatusManager;
 import org.sagebionetworks.repo.manager.table.ColumnModelManager;
 import org.sagebionetworks.repo.manager.table.TableRowManager;
-import org.sagebionetworks.repo.model.DatastoreException;
-import org.sagebionetworks.repo.model.UnauthorizedException;
-import org.sagebionetworks.repo.model.UserInfo;
 import org.sagebionetworks.repo.model.AuthorizationConstants.BOOTSTRAP_PRINCIPAL;
+import org.sagebionetworks.repo.model.DatastoreException;
+import org.sagebionetworks.repo.model.UserInfo;
 import org.sagebionetworks.repo.model.asynch.AsynchJobState;
 import org.sagebionetworks.repo.model.asynch.AsynchronousJobStatus;
 import org.sagebionetworks.repo.model.dao.FileHandleDao;
@@ -56,8 +57,9 @@ import com.amazonaws.services.s3.model.GetObjectRequest;
 @ContextConfiguration(locations = { "classpath:test-context.xml" })
 public class TableCSVDownloadWorkerIntegrationTest {
 
+	// This test can be slow when run from outside of Amazon.
+	public static final int MAX_WAIT_MS = 1000*60*3;
 	
-	public static final int MAX_WAIT_MS = 1000*60;
 	@Autowired
 	AsynchJobStatusManager asynchJobStatusManager;
 	@Autowired
@@ -88,6 +90,8 @@ public class TableCSVDownloadWorkerIntegrationTest {
 	
 	@Before
 	public void before() throws NotFoundException{
+		// Only run this test if the table feature is enabled.
+		Assume.assumeTrue(config.getTableEnabled());
 		semphoreManager.releaseAllLocksAsAdmin(new UserInfo(true));
 		// Start with an empty queue.
 		asynchJobStatusManager.emptyAllQueues();
@@ -98,19 +102,21 @@ public class TableCSVDownloadWorkerIntegrationTest {
 	
 	@After
 	public void after(){
-		if(adminUserInfo != null){
-			for(String id: toDelete){
-				try {
-					entityManager.deleteEntity(adminUserInfo, id);
-				} catch (Exception e) {}
+		if(config.getTableEnabled()){
+			if(adminUserInfo != null){
+				for(String id: toDelete){
+					try {
+						entityManager.deleteEntity(adminUserInfo, id);
+					} catch (Exception e) {}
+				}
 			}
-		}
-		if(tempFile != null){
-			tempFile.delete();
-		}
-		if(fileHandle != null){
-			s3Client.deleteObject(fileHandle.getBucketName(), fileHandle.getKey());
-			fileHandleDao.delete(fileHandle.getId());
+			if(tempFile != null){
+				tempFile.delete();
+			}
+			if(fileHandle != null){
+				s3Client.deleteObject(fileHandle.getBucketName(), fileHandle.getKey());
+				fileHandleDao.delete(fileHandle.getId());
+			}
 		}
 	}
